requirements.txt
fastapi
uvicorn
requests
pandas
numpy
python-dotenv
google-generativeai
groq
beautifulsoup4
httpx
apscheduler
